requirements.txt
Flask==2.2.5
mysql-connector-python==8.1.0
Flask-Login==0.6.2
Werkzeug==2.2.3
reportlab==4.0.0
python-dotenv==1.0.0
